chore(refactor): extract useEditorFlags into its own file
