+++ src/components/Assistant.vue
@@ -162,7 +162,8 @@ import NcActionButton from '@nextcloud/vue/components/NcActionButton'
 import NcActionSeparator from '@nextcloud/vue/components/NcActionSeparator'
 import NcListItem from '@nextcloud/vue/components/NcListItem'
 import NcModal from '@nextcloud/vue/components/NcModal'
-import { useEditorFlags, useFileMixin } from './Editor.provider.ts'
+import { useFileMixin } from './Editor.provider.ts'
+import { useEditorFlags } from '../composables/useEditorFlags.ts'
 import { useEditor } from '../composables/useEditor.ts'
 import { FloatingMenu } from '@tiptap/vue-2'
 import { emit, subscribe, unsubscribe } from '@nextcloud/event-bus'
