+++ src/components/CollisionResolveDialog.vue
@@ -29,7 +29,8 @@
 </template>
 
 <script>
-import { useEditorFlags, useSyncServiceMixin } from './Editor.provider.ts'
+import { useSyncServiceMixin } from './Editor.provider.ts'
+import { useEditorFlags } from '../composables/useEditorFlags.ts'
 import { useEditor } from '../composables/useEditor.ts'
 import NcButton from '@nextcloud/vue/components/NcButton'
 import { useEditorMethods } from '../composables/useEditorMethods.ts'
