+++ src/components/Editor.provider.ts
@@ -4,41 +4,6 @@
  */
 
 import { logger } from '../helpers/logger.js'
-import { provide, inject, ref, computed } from 'vue'
-import type { InjectionKey, Ref, ShallowReactive } from 'vue'
-import { isPublicShare } from '@nextcloud/sharing/public'
-import { loadState } from '@nextcloud/initial-state'
-
-export interface EditorFlags {
-	isPublic: Ref<boolean>
-	isRichEditor: Ref<boolean>
-	isRichWorkspace: Ref<boolean>
-}
-interface Props {
-	isDirectEditing: boolean
-	richWorkspace: boolean
-	mime: string
-}
-export const editorFlagsKey = Symbol('editor:flags') as InjectionKey<EditorFlags>
-export const provideEditorFlags = (props: ShallowReactive<Props>) => {
-	const isPublic = computed(() => props.isDirectEditing || isPublicShare())
-	const isRichWorkspace = computed(() => props.richWorkspace)
-	const isRichEditor = computed(
-		() =>
-			loadState('text', 'rich_editing_enabled', true)
-			&& props.mime === 'text/markdown',
-	)
-	provide(editorFlagsKey, { isPublic, isRichEditor, isRichWorkspace })
-	return { isPublic, isRichEditor, isRichWorkspace }
-}
-export const useEditorFlags = () => {
-	const { isPublic, isRichEditor, isRichWorkspace } = inject(editorFlagsKey, {
-		isPublic: ref(false),
-		isRichEditor: ref(false),
-		isRichWorkspace: ref(false),
-	})
-	return { isPublic, isRichEditor, isRichWorkspace }
-}
 
 export const FILE = Symbol('editor:file')
 export const ATTACHMENT_RESOLVER = Symbol('attachment:resolver')
